Added: Elasticsearch: Field query string support
