+++ elasticsearch-singledocument-plugin/README.md
@@ -1,0 +1,11 @@
+## Installing
+
+```bash
+bin/plugin install org.vertexium/vertexium-elasticsearch-singledocument-plugin/<version>
+```
+
+or
+
+```bash
+bin/plugin install file:///path/to/vertexium-elasticsearch-singledocument-plugin-<version>.zip
+```
