+++ elasticsearch-singledocument-plugin/src/main/resources/es-plugin.properties
@@ -1,0 +1,1 @@
+plugin=org.vertexium.elasticsearch.plugin.VertexiumElasticsearchPlugin
