+++ elasticsearch-singledocument/src/test/java/org/elasticsearch/common/logging/log4j/LogConfigurator.java
@@ -1,0 +1,10 @@
+package org.elasticsearch.common.logging.log4j;
+
+import org.elasticsearch.common.settings.Settings;
+
+// Override ES log configuration which overrides our logging
+public class LogConfigurator {
+    public static void configure(Settings settings) {
+
+    }
+}
