+++ elasticsearch-singledocument/src/test/resources/.gitignore
@@ -1,0 +1,1 @@
+vertexium-elasticsearch-singledocument-plugin.zip
