--- inmemory/src/main/java/org/vertexium/inmemory/security/package-info.java
@@ -1,5 +1,0 @@
-/**
- * These class are borrowed from Accumulo and Hadoop to provide the
- * visibility parsing and evaluation.
- */
-package org.vertexium.inmemory.security;
