+++ security/src/main/java/org/vertexium/security/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * These class are borrowed from Accumulo and Hadoop to provide the
+ * visibility parsing and evaluation.
+ */
+package org.vertexium.security;
